+++ scripts/variables.sh
@@ -1,1 +1,3 @@
 VAR_KEY_PREFIX="@sidebar-key"
+OPTION_DELIMITER="->"
+REGISTERED_PANE_PREFIX="@-sidebar-registered-pane"