+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 ### master
 - started the project
 - init script done
+- toggle script working